Separate things from Ui into Input, rename some functions
