+++ src/OpenLoco/src/Game.cpp
@@ -49,7 +49,7 @@ namespace OpenLoco::Game
         bool confirm = Ui::Windows::PromptBrowse::open(type, &_savePath[0], filter, titleId);
 
         Audio::unpauseSound();
-        Ui::processMessagesMini();
+        Input::processMessagesMini();
         SceneManager::unsetPauseFlag(1 << 2);
         Gfx::invalidateScreen();
         Gfx::renderAndUpdate();
