+++ src/OpenLoco/src/Input.cpp
@@ -7,12 +7,19 @@
 #include "Ui/Window.h"
 #include <map>
 
+#include <SDL2/SDL.h>
+#pragma warning(disable : 4121) // alignment of a member was sensitive to packing
+#include <SDL2/SDL_syswm.h>
+#pragma warning(default : 4121) // alignment of a member was sensitive to packing
+
 namespace OpenLoco::Input
 {
     static Flags _flags;
     static State _state;
     static Ui::Point32 _cursorDragStart;
     static uint32_t _cursorDragState;
+    static bool _exitRequested = false;
+    loco_global<uint8_t[256], 0x01140740> _keyboardState;
 
     void init()
     {
@@ -81,5 +88,189 @@ namespace OpenLoco::Input
         delta.y /= scale;
 
         return { static_cast<int16_t>(delta.x), static_cast<int16_t>(delta.y) };
+    }
+
+    // 0x004072EC
+    bool processMessagesMini()
+    {
+        SDL_Event e;
+        while (SDL_PollEvent(&e))
+        {
+            switch (e.type)
+            {
+                case SDL_QUIT:
+                    return false;
+                case SDL_WINDOWEVENT:
+                    switch (e.window.event)
+                    {
+                        case SDL_WINDOWEVENT_MOVED:
+                            Ui::windowPositionChanged(e.window.data1, e.window.data2);
+                            break;
+                        case SDL_WINDOWEVENT_SIZE_CHANGED:
+                            Ui::windowSizeChanged(e.window.data1, e.window.data2);
+                            break;
+                    }
+                    break;
+            }
+        }
+        return false;
+    }
+
+    // 0x0040477F
+    static void readKeyboardState()
+    {
+        addr<0x005251CC, uint8_t>() = 0;
+        auto dstSize = _keyboardState.size();
+        auto dst = _keyboardState.get();
+
+        int numKeys;
+
+        std::fill_n(dst, dstSize, 0);
+        auto keyboardState = SDL_GetKeyboardState(&numKeys);
+        if (keyboardState != nullptr)
+        {
+            for (int scanCode = 0; scanCode < numKeys; scanCode++)
+            {
+                bool isDown = keyboardState[scanCode] != 0;
+                if (!isDown)
+                {
+                    continue;
+                }
+
+                dst[scanCode] = 0x80;
+            }
+            addr<0x005251CC, uint8_t>() = 1;
+        }
+    }
+
+    // 0x00406FBA
+    static void handleKeyInput(uint32_t keycode)
+    {
+        Input::enqueueKey(keycode);
+
+        switch (keycode)
+        {
+            case SDLK_RETURN:
+            case SDLK_BACKSPACE:
+            case SDLK_DELETE:
+            {
+                char c[] = { (char)keycode, '\0' };
+                Input::enqueueText(c);
+                break;
+            }
+        }
+    }
+
+    // 0x0040726D
+    bool processMessages()
+    {
+        // The game has more than one loop for processing messages, if the secondary loop receives
+        // SDL_QUIT then the message would be lost for the primary loop so we have to keep track of it.
+        if (_exitRequested)
+        {
+            return false;
+        }
+
+        SDL_Event e;
+        while (SDL_PollEvent(&e))
+        {
+            switch (e.type)
+            {
+                case SDL_QUIT:
+                    _exitRequested = true;
+                    return false;
+                case SDL_WINDOWEVENT:
+                    switch (e.window.event)
+                    {
+                        case SDL_WINDOWEVENT_MOVED:
+                            Ui::windowPositionChanged(e.window.data1, e.window.data2);
+                            break;
+                        case SDL_WINDOWEVENT_SIZE_CHANGED:
+                            Ui::windowSizeChanged(e.window.data1, e.window.data2);
+                            break;
+                    }
+                    break;
+                case SDL_MOUSEMOTION:
+                {
+                    auto scaleFactor = Config::get().scaleFactor;
+                    auto x = static_cast<int32_t>(e.motion.x / scaleFactor);
+                    auto y = static_cast<int32_t>(e.motion.y / scaleFactor);
+                    auto xrel = static_cast<int32_t>(e.motion.xrel / scaleFactor);
+                    auto yrel = static_cast<int32_t>(e.motion.yrel / scaleFactor);
+                    Input::moveMouse(x, y, xrel, yrel);
+                    break;
+                }
+                case SDL_MOUSEWHEEL:
+                    Input::mouseWheel(e.wheel.y);
+                    break;
+                case SDL_MOUSEBUTTONDOWN:
+                {
+                    auto scaleFactor = Config::get().scaleFactor;
+                    const auto x = static_cast<int32_t>(e.button.x / scaleFactor);
+                    const auto y = static_cast<int32_t>(e.button.y / scaleFactor);
+                    addr<0x00525324, int32_t>() = 1;
+                    switch (e.button.button)
+                    {
+                        case SDL_BUTTON_LEFT:
+                            Input::enqueueMouseButton({ { x, y }, 1 });
+                            addr<0x0113E8A0, int32_t>() = 1;
+                            break;
+                        case SDL_BUTTON_RIGHT:
+                            Input::enqueueMouseButton({ { x, y }, 2 });
+                            addr<0x0113E0C0, int32_t>() = 1;
+                            setRightMouseButtonDown(true);
+                            addr<0x01140845, uint8_t>() = 0x80;
+                            break;
+                    }
+                    break;
+                }
+                case SDL_MOUSEBUTTONUP:
+                {
+                    auto scaleFactor = Config::get().scaleFactor;
+                    const auto x = static_cast<int32_t>(e.button.x / scaleFactor);
+                    const auto y = static_cast<int32_t>(e.button.y / scaleFactor);
+                    addr<0x00525324, int32_t>() = 1;
+                    switch (e.button.button)
+                    {
+                        case SDL_BUTTON_LEFT:
+                            Input::enqueueMouseButton({ { x, y }, 3 });
+                            addr<0x0113E8A0, int32_t>() = 0;
+                            break;
+                        case SDL_BUTTON_RIGHT:
+                            Input::enqueueMouseButton({ { x, y }, 4 });
+                            addr<0x0113E0C0, int32_t>() = 0;
+                            setRightMouseButtonDown(false);
+                            addr<0x01140845, uint8_t>() = 0;
+                            break;
+                    }
+                    break;
+                }
+                case SDL_KEYDOWN:
+                {
+                    auto keycode = e.key.keysym.sym;
+
+#if !(defined(__APPLE__) && defined(__MACH__))
+                    // Toggle fullscreen when ALT+RETURN is pressed
+                    if (keycode == SDLK_RETURN)
+                    {
+                        if ((e.key.keysym.mod & KMOD_LALT) || (e.key.keysym.mod & KMOD_RALT))
+                        {
+                            Ui::toggleFullscreenDesktop();
+                        }
+                    }
+#endif
+
+                    handleKeyInput(keycode);
+                    break;
+                }
+                case SDL_KEYUP:
+                    break;
+                case SDL_TEXTINPUT:
+                    enqueueText(e.text.text);
+                    break;
+            }
+        }
+        readKeyboardState();
+        return true;
     }
 }
