+++ src/OpenLoco/src/Ui.h
@@ -111,8 +111,6 @@ namespace OpenLoco::Ui
     void update();
     void triggerResize();
     void render();
-    bool processMessages();
-    bool processMessagesMini();
     void showMessageBox(const std::string& title, const std::string& message);
     Config::Resolution getResolution();
     Config::Resolution getDesktopResolution();
@@ -126,4 +124,11 @@ namespace OpenLoco::Ui
     void setWindowScaling(float newScaleFactor);
     void adjustWindowScale(float adjust_by);
     bool hasInputFocus();
+
+    void windowPositionChanged(int32_t x, int32_t y);
+    void windowSizeChanged(int32_t width, int32_t height);
+
+#if !(defined(__APPLE__) && defined(__MACH__))
+    void toggleFullscreenDesktop();
+#endif
 }
